# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Synthetic Solutions Inventory Edition API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      summary: List warehouses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [warehouses]
      summary: Create warehouse (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "201":
          description: Warehouse created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /warehouses/{id}:
    get:
      operationId: getWarehouse
      tags: [warehouses]
      summary: Get warehouse by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateWarehouse
      tags: [warehouses]
      summary: Update warehouse (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseUpdate"
      responses:
        "200":
          description: Updated warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteWarehouse
      tags: [warehouses]
      summary: Delete warehouse (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
  /warehouse-movements:
    get:
      operationId: listWarehouseMovements
      tags: [warehouses]
      summary: List warehouse stock movements
      security:
        - bearerAuth: []
      parameters:
        - name: warehouseId
          in: query
          schema:
            type: integer
        - name: productId
          in: query
          schema:
            type: integer
        - name: movementType
          in: query
          schema:
            type: string
            enum: [receipt, transfer, adjustment]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Warehouse stock movement ledger
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WarehouseMovement"
    post:
      operationId: createWarehouseMovement
      tags: [warehouses]
      summary: Record a warehouse receipt, transfer, or adjustment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseMovementInput"
      responses:
        "201":
          description: Movement recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseMovementResponse"
  /warehouse-transfers:
    post:
      operationId: createWarehouseTransfer
      tags: [warehouses]
      summary: Transfer stock from a warehouse to a store (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseTransferInput"
      responses:
        "201":
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseTransferResponse"
        "409":
          description: Insufficient stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /warehouses/{id}/movements:
    get:
      operationId: listWarehouseMovementsByWarehouse
      tags: [warehouses]
      summary: List movements for a warehouse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Warehouse stock movement ledger
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WarehouseMovement"
    post:
      operationId: createWarehouseMovementForWarehouse
      tags: [warehouses]
      summary: Record a movement for a warehouse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseMovementInput"
      responses:
        "201":
          description: Movement recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseMovementResponse"
  /warehouses/{id}/stock:
    get:
      operationId: getWarehouseStock
      tags: [warehouses]
      summary: Get current warehouse stock (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Current stock
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WarehouseStockLine"
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserProfile:
      type: object
      required: [id, name, role, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, store_user]
        storeId:
          type: ["integer", "null"]
        biometricEnabled:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # Store
...
    Warehouse:
      type: object
      required: [id, name, warehouseNumber, status, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        warehouseNumber:
          type: string
        manager:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zip:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, disabled]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    WarehouseInput:
      type: object
      required: [name, warehouseNumber]
      properties:
        name:
          type: string
        warehouseNumber:
          type: string
        manager:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: string
        email:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [active, inactive, disabled]
    WarehouseUpdate:
      type: object
      properties:
        name:
          type: string
        warehouseNumber:
          type: string
        manager:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: string
        email:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [active, inactive, disabled]
        isActive:
          type: boolean
    WarehouseMovement:
      type: object
      required: [id, warehouseId, productId, quantity, movementType, employeeId, movedAt]
      properties:
        id:
          type: integer
        warehouseId:
          type: integer
        productId:
          type: integer
        quantity:
          type: string
        movementType:
          type: string
          enum: [receipt, transfer, adjustment]
        transferGroupId:
          type: ["string", "null"]
        fromWarehouseId:
          type: ["integer", "null"]
        toWarehouseId:
          type: ["integer", "null"]
        destinationStoreId:
          type: ["integer", "null"]
        receivingRecordId:
          type: ["integer", "null"]
        employeeId:
          type: integer
        notes:
          type: ["string", "null"]
        movedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    WarehouseMovementInput:
      type: object
      required: [warehouseId, productId, quantity, movementType]
      properties:
        warehouseId:
          type: integer
        productId:
          type: integer
        quantity:
          type: string
        movementType:
          type: string
          enum: [receipt, transfer, adjustment]
        fromWarehouseId:
          type: integer
        toWarehouseId:
          type: integer
        receivingRecordId:
          type: integer
        notes:
          type: string
        movedAt:
          type: string
          format: date-time
    WarehouseMovementResponse:
      type: object
      properties:
        id:
          type: integer
        warehouseId:
          type: integer
        productId:
          type: integer
        quantity:
          type: string
        movementType:
          type: string
          enum: [receipt, transfer, adjustment]
        transferGroupId:
          type: string
        movements:
          type: array
          items:
            $ref: "#/components/schemas/WarehouseMovement"
    WarehouseTransferItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: string
    WarehouseTransferInput:
      type: object
      required: [sourceWarehouseId, destinationStoreId, items]
      properties:
        sourceWarehouseId:
          type: integer
        destinationStoreId:
          type: integer
        notes:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/WarehouseTransferItemInput"
    WarehouseTransferResponse:
      type: object
      required: [transferGroupId, header, receipt, movements]
      properties:
        transferGroupId:
          type: string
        header:
          type: object
          properties:
            id:
              type: integer
        receipt:
          type: object
          properties:
            id:
              type: integer
        movements:
          type: array
          items:
            $ref: "#/components/schemas/WarehouseMovement"
    WarehouseStockLine:
      type: object
      required: [warehouseId, productId, quantity]
      properties:
        warehouseId:
          type: integer
        productId:
          type: integer
        quantity:
          type: string
        lastCountDate:
          type: ["string", "null"]
          format: date-time

    # Users
...
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, store_user]
        storeId:
          type: ["integer", "null"]
        biometricEnabled:
          type: boolean
        isActive:
          type: boolean
...